Analyz/requirements.txt
Flask
flask-cors>=3.0.3
pandas>=2.2.0
python-dateutil>=2.9.0
numpy>=2.0.0
openpyxl>=3.1.2
requests>=2.31.0
python-dotenv>=1.0.1
Pillow>=10.4.0
psycopg2-binary>=2.9.9


